refactor: dissolve the shared Contracts namespace
Each interface moves to its real owner: Effect to Actions, ProvidesSchema
to Fragments, PageContract beside Page in Http, and the cross-cutting
Authorizable/DefinitionRegistry/HasLabel/DiscoversDefinitions into
Core\Contracts. The separate shared-contracts arch rule is no longer
needed since the kernel contracts now fall under the Core rule.
