+++ src/Actions/Components/Action.php
@@ -5,7 +5,7 @@ namespace Bambamboole\Lattice\Actions\Components;
 use BackedEnum;
 use Bambamboole\Lattice\Actions\ActionDefinition;
 use Bambamboole\Lattice\Actions\ActionRegistry;
-use Bambamboole\Lattice\Contracts\Effect;
+use Bambamboole\Lattice\Actions\Contracts\Effect;
 use Bambamboole\Lattice\Core\Components\Component;
 use Bambamboole\Lattice\Core\Components\IsInteractive;
