+++ src/Actions/Effect.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Bambamboole\Lattice\Actions;
 
+use Bambamboole\Lattice\Actions\Contracts\Effect as EffectContract;
 use Bambamboole\Lattice\Actions\Enums\EffectType;
-use Bambamboole\Lattice\Contracts\Effect as EffectContract;
 use Bambamboole\Lattice\Core\Enums\ToastType;
 use Bambamboole\Lattice\Core\Values\ToastMessage;
